Lint: satisfy ruff 0.16.1's default rules (executable bits, context managers, RUF046/059/100, PERF102)

+++ projects/tbox/tools/place_decoupling.py
@@ -37,7 +37,7 @@ def main():
     import pcbnew
 
     def mm(v):
-        return int(round(v * 1_000_000))
+        return round(v * 1_000_000)
 
     board = pcbnew.LoadBoard(PCB)
     fps = list(board.GetFootprints())
